Refactor CSS files in src/screens/OrganizationFundCampaign  (#2934)
* Update organization-fund-campaign css

* Fix linting

* Fix some issues

* Fix import issue

* Fix color change

* Update src/style/app.module.css


---------
