--- src/screens/OrganizationFundCampaign/OrganizationFundCampaign.module.css
@@ -1,202 +1,0 @@
-.organizationFundCampaignContainer {
-  margin: 0.5rem 0;
-}
-.goalButton {
-  border: 1px solid rgba(49, 187, 107, 1) !important;
-  color: rgba(49, 187, 107, 1) !important;
-  width: 75%;
-  padding: 10px;
-  border-radius: 8px;
-  display: block;
-  margin: auto;
-  box-shadow: 5px 5px 4px 0px rgba(49, 187, 107, 0.12);
-}
-.rowBackground {
-  background-color: var(--bs-white);
-  max-height: 120px;
-}
-.container {
-  min-height: 100vh;
-}
-.campaignModal {
-  max-width: 80vw;
-  margin-top: 2vh;
-  margin-left: 13vw;
-}
-.titlemodal {
-  color: #707070;
-  font-weight: 600;
-  font-size: 32px;
-  width: 65%;
-  margin-bottom: 0px;
-}
-.noOutline input {
-  outline: none;
-}
-.modalCloseBtn {
-  width: 40px;
-  height: 40px;
-  padding: 1rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.greenregbtn {
-  margin: 1rem 0 0;
-  margin-top: 15px;
-  border: 1px solid #e8e5e5;
-  box-shadow: 0 2px 2px #e8e5e5;
-  padding: 10px 10px;
-  border-radius: 5px;
-  background-color: #31bb6b;
-  width: 100%;
-  font-size: 16px;
-  color: white;
-  outline: none;
-  font-weight: 600;
-  cursor: pointer;
-  transition:
-    transform 0.2s,
-    box-shadow 0.2s;
-  width: 100%;
-  flex: 1;
-}
-
-.redregbtn {
-  margin: 1rem 0 0;
-  margin-top: 15px;
-  border: 1px solid #e8e5e5;
-  box-shadow: 0 2px 2px #e8e5e5;
-  padding: 10px 10px;
-  border-radius: 5px;
-  width: 100%;
-  font-size: 16px;
-  color: white;
-  outline: none;
-  font-weight: 600;
-  cursor: pointer;
-  transition:
-    transform 0.2s,
-    box-shadow 0.2s;
-  width: 100%;
-  flex: 1;
-}
-.campaignNameInfo {
-  font-size: medium;
-  cursor: pointer;
-}
-.campaignNameInfo:hover {
-  color: blue;
-  transform: translateY(-2px);
-}
-.message {
-  margin-top: 25%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.inputField {
-  background-color: white;
-  box-shadow: 0 1px 1px #31bb6b;
-}
-
-.dropdown {
-  background-color: white;
-  border: 1px solid #31bb6b;
-  position: relative;
-  display: inline-block;
-  color: #31bb6b;
-}
-
-.btnsContainer {
-  display: flex;
-  margin: 2rem 0 2rem 0;
-  gap: 0.8rem;
-}
-
-.btnsContainer .btnsBlock {
-  display: flex;
-  gap: 0.8rem;
-}
-
-.btnsContainer .btnsBlock div button {
-  display: flex;
-  margin-left: 1rem;
-  justify-content: center;
-  align-items: center;
-}
-
-.btnsContainer .input {
-  flex: 1;
-  position: relative;
-}
-
-.btnsContainer input {
-  outline: 1px solid var(--bs-gray-400);
-}
-
-.btnsContainer .input button {
-  width: 52px;
-}
-
-.mainpageright > hr {
-  margin-top: 20px;
-  width: 100%;
-  margin-left: -15px;
-  margin-right: -15px;
-  margin-bottom: 20px;
-}
-
-.tableHeader {
-  background-color: var(--bs-primary);
-  color: var(--bs-white);
-  font-size: 1rem;
-}
-
-@media (max-width: 1020px) {
-  .btnsContainer {
-    flex-direction: column;
-    margin: 1.5rem 0;
-  }
-
-  .btnsContainer .btnsBlock {
-    margin: 1.5rem 0 0 0;
-    justify-content: space-between;
-  }
-
-  .btnsContainer .btnsBlock div button {
-    margin: 0;
-  }
-
-  .createFundBtn {
-    margin-top: 0;
-  }
-}
-
-@media screen and (max-width: 575.5px) {
-  .mainpageright {
-    width: 98%;
-  }
-}
-
-/* For mobile devices  */
-
-@media (max-width: 520px) {
-  .btnsContainer {
-    margin-bottom: 0;
-  }
-
-  .btnsContainer .btnsBlock {
-    display: block;
-    margin-top: 1rem;
-    margin-right: 0;
-  }
-
-  .btnsContainer .btnsBlock div button {
-    margin-bottom: 1rem;
-    margin-right: 0;
-    width: 100%;
-  }
-}
